refactor(tabs): peel focus and group helpers out of tabHelpers
Move tabHelpers into a folder barrel so the first two clusters (session focus patches and tiled-group focus) can live in their own files without changing any import path.
